Added:Java functionality to no duplicate entry for the same person in a particular Address Book
