+++ src/com/bridgelabz_address_book/MultipleAddressBookHashMap.java
@@ -2,13 +2,14 @@ package com.bridgelabz_address_book;
 
 import java.sql.SQLOutput;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 import java.util.Scanner;
 
 public class MultipleAddressBookHashMap extends AddressBook {
 
     Scanner sc = new Scanner(System.in);
-     Map<String, AddressBook> addMultiple = new HashMap<>();
+
 
     static MultipleAddressBookHashMap multipleAddressBookHashMap = new MultipleAddressBookHashMap();
     AddressBook addressBook = new AddressBook();
@@ -36,6 +37,7 @@ public class MultipleAddressBookHashMap extends AddressBook {
                     System.out.println("Enter a name of address book to search");
                     String name = sc.next();
                     multipleAddressBookHashMap.showAddressBook(name);
+                    break;
                 case 5:
                     a = false;
                     break;
@@ -44,39 +46,20 @@ public class MultipleAddressBookHashMap extends AddressBook {
         } while (a);
     }
 
-    public void addAddressBook() {
-        System.out.println("Enter how many address book you want to add");
-        Scanner sc = new Scanner(System.in);
-        int numberOfBookAddress = sc.nextInt();
-        for (int i = 1; i <= numberOfBookAddress; i++) {
-            System.out.println("Entered " + i + " Address book name");
-            String addressBookName = sc.next();
-            addressBook.setAddressBookName(addressBookName);
-            addressBook.multiplePerson();
-            addMultiple.put(addressBookName, addressBook);
-//        addressBook.multiplyContact();
-//        addressBook.editDeleteContact();
-        }
-    }
-
     public  void showAddressBook(String name) {
-//        for (Map.Entry m:addMultiple.entrySet()) {
-//            System.out.println(m.getKey());
-//            AddressBook addressBook = entrySet.getValue();
-//            System.out.println("Details: " + addressBook.entrySet());
-//        }
-
-        for (Map.Entry<String, AddressBook> entry : addMultiple.entrySet()) {
-            System.out.println("Name: " + entry.getKey());
-
-            AddressBook addressBook = entry.getValue();
-            System.out.println( addressBook.showContacts()); // Replace with the actual method in your AddressBook class
-
-            // Print other details as needed
-            // System.out.println("SomeOtherDetail: " + addressBook.getSomeOtherDetail());
-
-            System.out.println(); // Add a newline for better readability
+        List<Contacts> addressBook = addMultiple.get(name);
+        for (Contacts c : addressBook) {
+            System.out.println(c.toString());
         }
+
+//        for (Map.Entry <String,List<Contacts>>m:addMultiple.entrySet()) {
+//            System.out.println(m.getKey());
+//            List<Contacts> addressBook = m.getValue();
+//            for(Contacts c:addressBook)
+//            {
+//                System.out.println(c.toString());
+//            }
+//        }
 
     }
 }
